+++ Rendering/WebGPU/Private/vtkWebGPUComputePassTextureStorageInternals.h
@@ -246,7 +246,8 @@ public:
     const std::string textureLabel = texture->GetLabel();
     const char* textureLabelCStr = textureLabel.c_str();
     const uint32_t bytesPerRow = texture->GetBytesPerPixel() * texture->GetWidth();
-    this->ParentPassWGPUConfiguration->WriteTexture(wgpuTexture.Get(), bytesPerRow, numBytes, bytes,
+    this->ParentPassWGPUConfiguration->WriteTexture(wgpuTexture.Get(), bytesPerRow,
+      static_cast<uint32_t>(numBytes), bytes,
       /*srcOffset=*/0, /*dstOffset=*/{ 0, 0, 0 }, /*dstMipLevel=*/0,
       /*textureLabel=*/textureLabelCStr);
   }
